+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (unreleased)
 
+## 1.1.0
+
 - Annotate NoMethodError in non-production environments (https://github.com/zombocom/dead_end/pull/46)
 - Do not count trailing if/unless as a keyword (https://github.com/zombocom/dead_end/pull/44)